requirements.txt
asgiref==3.2.3
certifi==2019.11.28
Django==3.0.3
django-ckeditor==5.9.0
django-js-asset==1.2.2
django-taggit==1.2.0
mysql-connector==2.2.9
Pillow==7.0.0
pytz==2019.3
sqlparse==0.3.0
wincertstore==0.2
